+++ src/streamIO/src/streamWrite.c
@@ -44,7 +44,7 @@ errorCode writeNextBit(EXIStream* strm, boolean bit_val)
 	return EXIP_OK;
 }
 
-errorCode writeNBits(EXIStream* strm, unsigned char nbits, unsigned int bits_val)
+errorCode writeNBits(EXIStream* strm, unsigned char nbits, unsigned long bits_val)
 {
 	unsigned int numBitsWrite = 0; // Number of the bits written so far
 	unsigned char tmp = 0;
